Fix Sonar violations

+++ src/main/java/com/sivalabs/devzone/web/api/resources/UserRestController.java
@@ -9,9 +9,18 @@ import com.sivalabs.devzone.domain.services.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
+import java.util.Optional;
 
 import static org.springframework.http.HttpStatus.CREATED;
 
@@ -25,7 +34,7 @@ public class UserRestController {
 
     @GetMapping("/{id}")
     public ResponseEntity<UserDTO> getUser(@PathVariable Long id) {
-        log.info("process=get_user, user_id=" + id);
+        log.info("process=get_user, user_id={}", id);
         return userService.getUserById(id)
             .map(ResponseEntity::ok)
             .orElse(ResponseEntity.notFound().build());
@@ -34,7 +43,7 @@ public class UserRestController {
     @PostMapping("")
     @ResponseStatus(CREATED)
     public UserDTO createUser(@RequestBody @Valid CreateUserRequest createUserRequest) {
-        log.info("process=create_user, user_email=" + createUserRequest.getEmail());
+        log.info("process=create_user, user_email={}", createUserRequest.getEmail());
         UserDTO userDTO = new UserDTO(
             null,
             createUserRequest.getName(),
@@ -49,7 +58,7 @@ public class UserRestController {
     @PutMapping("/{id}")
     @AnyAuthenticatedUser
     public UserDTO updateUser(@PathVariable Long id, @RequestBody @Valid UserDTO user) {
-        log.info("process=update_user, user_id=" + id);
+        log.info("process=update_user, user_id={}", id);
         if (!id.equals(securityService.loginUserId())) {
             throw new ResourceNotFoundException("User not found with id=" + id);
         }
@@ -59,16 +68,16 @@ public class UserRestController {
 
     @DeleteMapping("/{id}")
     @AnyAuthenticatedUser
-    public void deleteUser(@PathVariable Long id) {
-        log.info("process=delete_user, user_id=" + id);
-        userService.getUserById(id).map(u -> {
-            if (!id.equals(securityService.loginUserId()) && !securityService.isCurrentUserAdmin()) {
-                throw new ResourceNotFoundException("User not found with id=" + id);
-            }
-            userService.deleteUser(id);
-            return u;
-        });
+    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+        log.info("process=delete_user, user_id={}", id);
+        Optional<UserDTO> userById = userService.getUserById(id);
+        if (userById.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        if (!id.equals(securityService.loginUserId()) && !securityService.isCurrentUserAdmin()) {
+            return ResponseEntity.notFound().build();
+        }
+        userService.deleteUser(id);
+        return ResponseEntity.ok().build();
     }
-
-
 }
